+++ ScripBaseDeDatos/Backend/PuraVidaStoreBK/PuraVidaStoreBK/ExecQuerys/Interfaces/IProductoQuery.cs
@@ -8,7 +8,7 @@ namespace PuraVidaStoreBK.ExecQuerys.Interfaces
         public Task<List<Producto>> ListaProductosFiltrada();
         public Task<List<Producto>> ListaProductosNoFiltrada();
         public Task<List<Producto>> ListaProductos();
-        public Task<Producto> ProductoPorId(int id);
+        public Task<Producto> ProductoPorId(long id);
         public Task<Producto> BuscarProductoPorCodigo(string codigo);
         public Task<List<Producto>> ProductoPorDescripcion(string Descripcion);
         public Task<List<Producto>> ProductoPorNoFiltradaDescripcion(string Descripcion);
