+++ ScripBaseDeDatos/Backend/PuraVidaStoreBK/PuraVidaStoreBK/Models/DTOS/PersonaDto.cs
@@ -4,10 +4,14 @@ namespace PuraVidaStoreBK.Models.DTOS
 {
     public class PersonaDto
     {
-        public int PsrId { get; set; }
+        public long PsrId { get; set; }
+
         public string PsrIdentificacion { get; set; } = null!;
+
         public string PsrNombre { get; set; } = null!;
+
         public string PsrApellido1 { get; set; } = null!;
+
         public string PsrApellido2 { get; set; } = null!;
 
     }
